Refactor: move composable related to update in subfolder
