+++ app/src/main/java/earth/mp3player/ui/components/settings/updates/CheckUpdateButton.kt
@@ -23,56 +23,46 @@
  * PS: I don't answer quickly.
  */
 
-package earth.mp3player.ui.components.settings
+package earth.mp3player.ui.components.settings.updates
 
 import android.content.Context
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.hapticfeedback.HapticFeedback
+import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import earth.mp3player.internet.R
-import earth.mp3player.internet.UpdateAvailableStatus
+import earth.mp3player.R
 import earth.mp3player.internet.UpdateManager
-import earth.mp3player.ui.components.settings.utils.openUrl
 
 /**
- * @author Antoine Pirlot on 14/04/2024
+ * @author Antoine Pirlot on 11/04/2024
  */
 
 @Composable
-fun UpdateAvailable(
+fun CheckUpdateButton(
     modifier: Modifier = Modifier,
 ) {
     val context: Context = LocalContext.current
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
+    val haptics: HapticFeedback = LocalHapticFeedback.current
+    Button(
+        modifier = modifier,
+        onClick = {
+            haptics.performHapticFeedback(hapticFeedbackType = HapticFeedbackType.TextHandleMove)
+            UpdateManager.checkUpdate(context = context)
+        }
     ) {
-        Text(text = stringResource(id = R.string.update_available))
-        Button(onClick = {
-            openUrl(
-                context = context,
-                url = UpdateAvailableStatus.AVAILABLE.updateLink!!
-            )
-        }) {
-            Text(text = stringResource(id = R.string.see_details))
-        }
-        Button(onClick = { UpdateManager.downloadUpdateApk(context = context) }) {
-            Text(text = stringResource(id = R.string.download_update))
-        }
+        Text(text = stringResource(id = R.string.check_update))
+
     }
 }
 
 @Preview
 @Composable
-fun UpdateAvailablePreview() {
-    UpdateAvailable()
+fun CheckUpdateButtonPreview() {
+    CheckUpdateButton()
 }
